Add files via upload
Our final major project at college involved the development of an automated attendance system using deep learning. We used advanced machine learning and deep learning algorithms to create a system that could accurately track and record attendance in real-time. The system uses face recognition technology to detect and recognize the faces of students while they are attending lectures, making it easy to use and efficient. This project has the potential to revolutionize the way attendance is recorded in educational institutions, saving time and reducing errors.